(BUGFIX) Fix the auto selection when "}" or "]" are typed
When typing ')', '"', '>', '}' or ']' sn is supposed to select
the contents between the parenthesis, the braces or whatever.

If you need the right alt key to get to the braces and brackets
sn wouldn't do it. Now it does, but I had to make sn check every
time you press a key while in the text widget, hopefully it won't
be to slow, doesn't look like it is but yell if it is for you.